chore: extracted NavigationList from NavigationBarDesktop and NavigationBarMobile
