+++ src/layouts/DashboardLayout/components/NavigationBar/NavigationBarMobile.jsx
@@ -1,42 +1,10 @@
-import { Fragment, useEffect, useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import { Fragment } from "react";
 import { Dialog, Transition } from "@headlessui/react";
-import { XIcon, LogoutIcon } from "@heroicons/react/outline";
+import { XIcon } from "@heroicons/react/outline";
 
-import dashboard from "~/config/dashboard";
-import { selectTeam } from "~/store/actions/teamActions";
-
-import BaseCombobox from "~/components/generic/form/BaseCombobox";
-import NavigationBarItem from "./NavigationBarItem";
+import NavigationList from "./NavigationList";
 
 function NavigationBarMobile({ sidebarOpen, setSidebarOpen }) {
-  const [query, setQuery] = useState("");
-
-  const [selectedTeam, setSelectedTeam] = useState({});
-
-  const dispatch = useDispatch();
-
-  const { data: acceptedTeams } = useSelector((state) => state.acceptedTeams);
-  const selectedTeamId = useSelector((state) => state.selectedTeamId);
-  const { data: notification } = useSelector((state) => state.notification);
-
-  const filteredTeams =
-    query === ""
-      ? acceptedTeams
-      : acceptedTeams.filter((team) =>
-          team.name.toLowerCase().includes(query.toLowerCase()),
-        );
-
-  useEffect(() => {
-    setSelectedTeam(acceptedTeams.find(({ _id }) => _id === selectedTeamId));
-  }, [selectedTeamId, acceptedTeams]);
-
-  const handleSelectTeam = (team) => {
-    dispatch(selectTeam(team));
-  };
-
-  const handleLogout = () => {};
-
   return (
     <Transition.Root show={sidebarOpen} as={Fragment}>
       <Dialog
@@ -93,43 +61,7 @@ function NavigationBarMobile({ sidebarOpen, setSidebarOpen }) {
               />
             </div>
             <div className="mt-5 h-0 flex-1 overflow-y-auto">
-              <nav className="space-y-1 px-2">
-                {acceptedTeams.length > 0 && (
-                  <BaseCombobox
-                    className="mb-4"
-                    value={selectedTeam}
-                    onChange={handleSelectTeam}
-                    filteredItem={filteredTeams}
-                    setQuery={setQuery}
-                  />
-                )}
-                {acceptedTeams.length > 0 ? (
-                  Object.values(dashboard).map((navigation) => {
-                    return navigation.path === "/discussion" ? (
-                      <NavigationBarItem
-                        {...navigation}
-                        key={navigation.name}
-                        notification={notification}
-                      />
-                    ) : (
-                      <NavigationBarItem
-                        {...navigation}
-                        key={navigation.name}
-                      />
-                    );
-                  })
-                ) : (
-                  <NavigationBarItem {...dashboard.proposal} />
-                )}
-                <div className="pt-3">
-                  <NavigationBarItem
-                    name="Keluar"
-                    path="#"
-                    icon={LogoutIcon}
-                    onLogout={handleLogout}
-                  />
-                </div>
-              </nav>
+              <NavigationList />
             </div>
           </div>
         </Transition.Child>
